Be 04 김수빈

+++ backend/src/main/java/com/web/curation/service/TalkService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -44,9 +45,41 @@ public class TalkService {
         return getTalkDtos(listTalk);
     }
 
-    public List<TalkDto> userTalk(int userId) {
-        List<Talk> listTalk = talkRepository.findByUser(userRepository.findByUserId(userId));
+    public List<TalkDto> userTalk(String email) {
+        List<Talk> listTalk = talkRepository.findByUser(userRepository.getByEmail(email));
         return getTalkDtos(listTalk);
+    }
+
+    public TalkDto detailTalk(int talkId) {
+        Talk talk = talkRepository.findByTalkId(talkId);
+        // 상세 조회 시 조회 수 증가
+        int cnt = talk.getClick();
+        talk.setClick(cnt+1);
+        talkRepository.save(talk);
+
+        TalkDto talkDto = new TalkDto();
+
+        talkDto.setProfileImgPath(talk.getUser().getProfileImg());
+
+        talkDto.setTalkId(talk.getTalkId());
+        talkDto.setNickname(talk.getUser().getNickname());
+        talkDto.setTitle(talk.getTitle());
+        talkDto.setContents(talk.getContents());
+        talkDto.setHashtag(talk.getHashtag());
+        talkDto.setUploadDate(talk.getUploadDate());
+
+        // 좋아요 수
+        int countLike = talkLikeRepository.countByTalk(talk);
+        talkDto.setLike(countLike);
+
+        // 조회수
+        talkDto.setClick(talk.getClick());
+
+        ThumbnailFile thumbnailFile = thumbnailFileRepository.findByTalk(talk);
+        talkDto.setThumbnailFileName(thumbnailFile.getName());
+        talkDto.setThumbnailFilePath(thumbnailFile.getFilePath());
+
+        return talkDto;
     }
 
 
@@ -61,26 +94,30 @@ public class TalkService {
             talkDto.setNickname(talk.getUser().getNickname());
             talkDto.setUploadDate(talk.getUploadDate());
             talkDto.setTitle(talk.getTitle());
-            talkDto.setThumbnailFileName(talk.getThumbnailFile().getName());
-            talkDto.setThumbnailFilePath(talk.getThumbnailFile().getFilePath());
+//            talkDto.setThumbnailFileName(talk.get().getName());
+//            talkDto.setThumbnailFilePath(talk.getThumbnailFile().getFilePath());
             talkDto.setHashtag(talk.getHashtag());
 
             // 좋아요 수
-            Long countLike = talkLikeRepository.countByTalk(talk);
+            int countLike = talkLikeRepository.countByTalk(talk);
             talkDto.setLike(countLike);
 
             // 조회수
             talkDto.setClick(talk.getClick());
 
-            talkDto.setTalkContents(talk.getTalkContents());
+            talkDto.setContents(talk.getContents());
+
+            ThumbnailFile thumbnailFile = thumbnailFileRepository.findByTalk(talk);
+            talkDto.setThumbnailFilePath(thumbnailFile.getFilePath());
+            talkDto.setThumbnailFileName(thumbnailFile.getName());
 
             listTalk.add(talkDto);
         }
 
         return listTalk;
     }
-
-    public boolean writeTalk(TalkDto talkDto) {
+    @Transactional
+    public int writeTalk(TalkDto talkDto) {
         LOGGER.info("[talk 게시글 등록] 게시글 제목 : {}", talkDto.getTitle());
         // talk 저장
         Talk talk = new Talk();
@@ -102,29 +139,65 @@ public class TalkService {
 
         thumbnailFileRepository.save(thumbnailFile);
 
-
-        // TalkContent 저장 받아오는거 따라서 만들어야함.
-
+        talk.setContents(talkDto.getContents());
 
         LOGGER.info("[getSignUpResult] userEntity 값이 들어왔는지 확인 후 결과값 주입");
         if (savedTalk != null) {
             LOGGER.info("talk 게시글 저장 완료");
-            return true;
+            int talkId = savedTalk.getTalkId();
+            return talkId;
         } else {
             LOGGER.info("talk 게시글 저장 실패");
-            return false;
+            return 0;
         }
     }
 
+    public boolean updateTalk(TalkDto talkDto) {
 
+        Talk talk = talkRepository.findByTalkId(talkDto.getTalkId());
 
+        talk.setTitle(talkDto.getTitle());
+        talk.setContents(talkDto.getContents());
+        talk.setHashtag(talkDto.getHashtag());
 
+        talkRepository.save(talk);
 
+        // file 이미지 변경하기
+        ThumbnailFile thumbnailFile = thumbnailFileRepository.findByTalk(talk);
 
-    public boolean pushLike(int talkId, String email) {
+        thumbnailFile.setTalk(talk);
+        thumbnailFile.setName(talkDto.getThumbnailFileName());
+        thumbnailFile.setFilePath(talkDto.getThumbnailFilePath());
+
+        thumbnailFileRepository.save(thumbnailFile);
+
+        if(talk.getTitle().equals(talkDto.getTitle()) && talk.getHashtag().equals(talkDto.getHashtag())){
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteTalk(int talkId) {
+        // 게시글 삭제
+        Talk talk = talkRepository.findByTalkId(talkId);
+        talkRepository.delete(talk);
+        // 사진 삭제
+        ThumbnailFile thumbnailFile = thumbnailFileRepository.findByTalk(talk);
+        thumbnailFileRepository.delete(thumbnailFile);
+
+        Talk verify = talkRepository.findByTalkId(talkId);
+
+        if(verify == null) return true;
+        return false;
+    }
+
+    public int pushLike(int talkId, String email) {
         Talk talk = talkRepository.findByTalkId(talkId);
 
         User user = userRepository.getByEmail(email);
+
+        // 좋아요를 두번 누르면 fail
+        if(talkLikeRepository.existsByUserAndTalk(user, talk)) return -1;
 
         TalkLike talkLike = new TalkLike();
         talkLike.setTalk(talk);
@@ -132,29 +205,37 @@ public class TalkService {
 
         talkLikeRepository.save(talkLike);
 
-        TalkLike savedTalkLike = talkLikeRepository.findByLikeId(talkLike.getLikeId());
+//        TalkLike savedTalkLike = talkLikeRepository.findByLikeId(talkLike.getLikeId());
+        // 좋아요 수
+        int countLike = talkLikeRepository.countByTalk(talk);
 
-        if (savedTalkLike != null) {
+        if (talkLikeRepository.existsByLikeId(talkLike.getLikeId())) {
             LOGGER.info("push like 완료");
-            return true;
+            return countLike;
         } else {
             LOGGER.info("push like 실패");
-            return false;
+            return -1;
         }
     }
 
 
-    public boolean cancelLike(int talkId, String email) {
+    public int cancelLike(int talkId, String email) {
 
         Talk talk = talkRepository.findByTalkId(talkId);
         User user = userRepository.getByEmail(email);
 
         TalkLike talkLike = talkLikeRepository.findByTalkAndUser(talk, user);
 
+        // 좋아요 수
+        int countLike = talkLikeRepository.countByTalk(talk);
+
+        // 좋아요 수가 1이면 fail
+        if(countLike == 0) return -1;
+
         talkLikeRepository.delete(talkLike);
 
-        if(!talkLikeRepository.existsByLikeId(talkLike.getLikeId())) return true;
-        return false;
+        if(!talkLikeRepository.existsByLikeId(talkLike.getLikeId())) return countLike-1;
+        return -1;
     }
